+++ src/components/FieldOrganization/FieldOrganization.test.js
@@ -3,7 +3,7 @@ import {
   act,
   waitFor,
   fireEvent,
-} from '@testing-library/react';
+} from '@folio/jest-config-stripes/testing-library/react';
 import { runAxeTest } from '@folio/stripes-testing';
 
 import '../../../test/jest/__mock__';