+++ src/components/Jobs/Jobs.test.js
@@ -3,7 +3,7 @@ import {
   fireEvent,
   waitFor,
   within,
-} from '@testing-library/react';
+} from '@folio/jest-config-stripes/testing-library/react';
 import { runAxeTest } from '@folio/stripes-testing';
 
 import {