+++ src/components/ProfileTree/ProfileTree.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {
   fireEvent,
   waitFor,
-} from '@testing-library/react';
+} from '@folio/jest-config-stripes/testing-library/react';
 import { noop } from 'lodash';
 import { runAxeTest } from '@folio/stripes-testing';