+++ src/components/SearchAndSort/SearchAndSort.test.js
@@ -7,7 +7,7 @@ import {
   fireEvent,
   waitFor,
   render,
-} from '@testing-library/react';
+} from '@folio/jest-config-stripes/testing-library/react';
 import { noop } from 'lodash';
 import { runAxeTest } from '@folio/stripes-testing';