+++ src/components/TextDate/TextDate.test.js
@@ -1,5 +1,5 @@
 import React, { createRef } from 'react';
-import { fireEvent } from '@testing-library/react';
+import { fireEvent } from '@folio/jest-config-stripes/testing-library/react';
 import { noop } from 'lodash';
 import { runAxeTest } from '@folio/stripes-testing';