+++ src/settings/MappingProfiles/detailsSections/hooks/useFieldMappingBoolFieldValue.test.js
@@ -1,4 +1,4 @@
-import { renderHook } from '@testing-library/react';
+import { renderHook } from '@folio/jest-config-stripes/testing-library/react';
 import * as reactRedux from 'react-redux';
 
 import '../../../../../test/jest/__mock__';